Decompose and document the table components
