+++ src/component-library/Table/FilterParameters/DefaultTextFilterParams.ts
@@ -1,0 +1,7 @@
+import {ITextFilterParams} from "ag-grid-community";
+
+export const DefaultTextFilterParams: ITextFilterParams = {
+    filterOptions: ['contains'],
+    buttons: ['reset'],
+    maxNumConditions: 1,
+};
