+++ src/component-library/Table/Overlays/NoDataYetOverlay.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const NoDataYetOverlay = () => {
+    return <div className="text-neutral-100">Click <b>Search</b></div>
+};
